requirements.txt
boto3==1.34.0
Pillow==10.2.0
numpy==1.26.3
rembg==2.0.67
onnxruntime==1.17.0
opencv-python-headless==4.11.0.86
scipy==1.11.4
mediapipe==0.10.21